refactor: break down class
